Use TypeScript to validate the various fields that apply to Courses, Instructors, and Students
Use TypeScript to validate the various fields that apply to Courses, Instructors, and Students. You have some latitude in determining what the properties of these items will be but you should ensure that nothing invalid can pass from the web interface into the back-end.

If the user can enter something that ends up throwing an error on the back-end, then it should have been caught by the validation performed with TypeScript on the front end.